// src/content/engineUnitFields.generated.ts
export const ENGINE_OPEN_TABLES: { unit: string[]; weapon: string[] } = {
  unit: ["SFXTypes.*", "buildOptions", "customParams"],
  weapon: ["customParams", "damage", "scarIndices"],
};